feat: modularised ping
License: MIT
